[all] Add `Error` raw action to represent invalid actions
This commit adds a new _virtual_ raw action (like `Unknown`) to allow the representation of invalid actions. Unlike `Unknown` where the parsers and emitters do not attempt to understand the action, `Error` indicates that the action is _known_ to be invalid and would trigger a "corrupted data" error in Adobe's Flash Player. The optional `error` field can provide additional information.