[all] Add `End` raw action
This commit adds the raw `End` action as defined is the SWF specification. This action was omitted initially because of the wrong assumption that the raw actions could be represent as a list with an implicit final end action. The ability to jump to arbitrary positions in the AVM1 byte code invalidated this assumption. This forced parsers to add a special case for the `End` action. With this commit, it will be possible to remove this special case and simplify downstream implementations.