[all] Rename project to `avm1-types`
This commit updates the project name to the more fitting `avm1-types` instead of `avm1-tree`. The project was originally born in `swf-tree` and used a "tree of actions" representation. This representation was then scrapped in favor of the two levels: raw action and Control Flow Graph (CFG). The name `avm1-types` indicates more clearly the purpose of this library and avoid connections to the legacy tree representation.